Revert "Cache change feeds durably on blob workers (cherry pick #10460) (#10499)" (#10536)
This reverts commit 53229482b347a93a899efa7adf777be09d059ad0.
