+++ fdbclient/NativeAPI.actor.cpp
@@ -9436,104 +9436,6 @@ ACTOR Future<Void> storageFeedVersionUpdater(StorageServerInterface interf, Chan
 	}
 }
 
-ACTOR Future<Void> changeFeedCommitter(IKeyValueStore* storage,
-                                       Reference<AsyncVar<bool>> commitChangeFeedStorage,
-                                       int64_t* uncommittedCFBytes) {
-	loop {
-		while (!commitChangeFeedStorage->get()) {
-			wait(commitChangeFeedStorage->onChange());
-		}
-		*uncommittedCFBytes = 0;
-		commitChangeFeedStorage->set(false);
-		wait(storage->commit());
-	}
-}
-
-ACTOR Future<Void> cleanupChangeFeedCache(DatabaseContext* db) {
-	wait(db->initializeChangeFeedCache);
-	wait(delay(CLIENT_KNOBS->CHANGE_FEED_CACHE_EXPIRE_TIME));
-	loop {
-		for (auto it = db->changeFeedCaches.begin(); it != db->changeFeedCaches.end(); ++it) {
-			if (!it->second->active && now() - it->second->inactiveTime > CLIENT_KNOBS->CHANGE_FEED_CACHE_EXPIRE_TIME) {
-				Key beginKey = changeFeedCacheKey(it->first.tenantPrefix, it->first.rangeId, it->first.range, 0);
-				Key endKey =
-				    changeFeedCacheKey(it->first.tenantPrefix, it->first.rangeId, it->first.range, MAX_VERSION);
-				db->storage->clear(KeyRangeRef(beginKey, endKey));
-				KeyRange feedRange =
-				    singleKeyRange(changeFeedCacheFeedKey(it->first.tenantPrefix, it->first.rangeId, it->first.range));
-				db->storage->clear(feedRange);
-
-				db->uncommittedCFBytes += beginKey.size() + endKey.size() + feedRange.expectedSize();
-				if (db->uncommittedCFBytes > CLIENT_KNOBS->CHANGE_FEED_CACHE_FLUSH_BYTES) {
-					db->commitChangeFeedStorage->set(true);
-				}
-
-				auto& rangeIdCache = db->rangeId_cacheData[it->first.rangeId];
-				rangeIdCache.erase(it->first);
-				if (rangeIdCache.empty()) {
-					db->rangeId_cacheData.erase(it->first.rangeId);
-				}
-				db->changeFeedCaches.erase(it);
-				break;
-			}
-		}
-		wait(delay(5.0));
-	}
-}
-
-ACTOR Future<Void> initializeCFCache(DatabaseContext* db) {
-	state Key beginKey = changeFeedCacheFeedKeys.begin;
-	loop {
-		RangeResult res = wait(db->storage->readRange(KeyRangeRef(beginKey, changeFeedCacheFeedKeys.end),
-		                                              CLIENT_KNOBS->CHANGE_FEED_CACHE_LIMIT_BYTES,
-		                                              CLIENT_KNOBS->CHANGE_FEED_CACHE_LIMIT_BYTES));
-		if (res.size()) {
-			beginKey = keyAfter(res.back().key);
-		} else {
-			ASSERT(!res.more);
-		}
-		for (auto& kv : res) {
-			ChangeFeedCacheRange cf(decodeChangeFeedCacheFeedKey(kv.key));
-			Reference<ChangeFeedCacheData> data = makeReference<ChangeFeedCacheData>();
-			auto val = decodeChangeFeedCacheFeedValue(kv.value);
-			data->version = val.first;
-			data->popped = val.second;
-			data->active = false;
-			data->inactiveTime = now();
-			db->changeFeedCaches[cf] = data;
-			db->rangeId_cacheData[cf.rangeId][cf] = data;
-		}
-		if (!res.more) {
-			break;
-		}
-	}
-	return Void();
-}
-
-ACTOR Future<Void> handleShutdown(DatabaseContext* db) {
-	try {
-		wait(db->storage->getError());
-	} catch (Error& e) {
-		TraceEvent("ChangeFeedCacheDiskError").error(e);
-	}
-	db->initializeChangeFeedCache = Void();
-	db->storage = nullptr;
-	db->changeFeedStorageCommitter = Void();
-	return Void();
-}
-
-void DatabaseContext::setStorage(IKeyValueStore* store) {
-	if (storage != nullptr) {
-		TraceEvent(SevError, "NativeClientMultipleSetStorage");
-		return;
-	}
-	storage = store;
-	commitChangeFeedStorage = makeReference<AsyncVar<bool>>(false);
-	initializeChangeFeedCache = initializeCFCache(this);
-	changeFeedStorageCommitter = changeFeedCommitter(storage, commitChangeFeedStorage, &uncommittedCFBytes) &&
-	                             cleanupChangeFeedCache(this) && handleShutdown(this);
-}
-
 Reference<ChangeFeedStorageData> DatabaseContext::getStorageData(StorageServerInterface interf) {
 	// use token from interface since that changes on SS restart
 	UID token = interf.waitFailure.getEndpoint().token;
@@ -10113,42 +10015,12 @@ ACTOR Future<Void> partialChangeFeedStream(StorageServerInterface interf,
 	}
 }
 
-void writeMutationsToCache(Reference<ChangeFeedCacheData> cacheData,
-                           Reference<DatabaseContext> db,
-                           Standalone<VectorRef<MutationsAndVersionRef>> cacheOut,
-                           Key rangeID,
-                           KeyRange range,
-                           Key tenantPrefix) {
-	if (!cacheData) {
-		return;
-	}
-	ASSERT(cacheData->active);
-	while (!cacheOut.empty() && cacheOut.front().version <= cacheData->latest) {
-		cacheOut.pop_front();
-	}
-	if (!cacheOut.empty()) {
-		Key durableKey = changeFeedCacheKey(tenantPrefix, rangeID, range, cacheOut.back().version);
-		Value durableValue = changeFeedCacheValue(cacheOut);
-		db->storage->set(KeyValueRef(durableKey, durableValue));
-		cacheData->latest = cacheOut.back().version;
-		db->uncommittedCFBytes += durableKey.size() + durableValue.size();
-		if (db->uncommittedCFBytes > CLIENT_KNOBS->CHANGE_FEED_CACHE_FLUSH_BYTES) {
-			db->commitChangeFeedStorage->set(true);
-		}
-	}
-}
-
 ACTOR Future<Void> mergeChangeFeedStreamInternal(Reference<ChangeFeedData> results,
-                                                 Key rangeID,
-                                                 KeyRange range,
                                                  std::vector<std::pair<StorageServerInterface, KeyRange>> interfs,
                                                  std::vector<MutationAndVersionStream> streams,
                                                  Version* begin,
                                                  Version end,
-                                                 UID mergeCursorUID,
-                                                 Reference<DatabaseContext> db,
-                                                 Reference<ChangeFeedCacheData> cacheData,
-                                                 Key tenantPrefix) {
+                                                 UID mergeCursorUID) {
 	state Promise<Void> refresh = results->refresh;
 	// with empty version handling in the partial cursor, all streams will always have a next element with version >=
 	// the minimum version of any stream's next element
@@ -10245,7 +10117,7 @@ ACTOR Future<Void> mergeChangeFeedStreamInternal(Reference<ChangeFeedData> resul
 			ASSERT(results->mutations.isEmpty());
 		} else {
 			ASSERT(nextOut.back().version > results->lastReturnedVersion.get());
-			writeMutationsToCache(cacheData, db, nextOut, rangeID, range, tenantPrefix);
+
 			results->mutations.send(nextOut);
 			wait(results->mutations.onEmpty());
 			wait(delay(0));
@@ -10261,15 +10133,12 @@ ACTOR Future<Void> mergeChangeFeedStream(Reference<DatabaseContext> db,
                                          std::vector<std::pair<StorageServerInterface, KeyRange>> interfs,
                                          Reference<ChangeFeedData> results,
                                          Key rangeID,
-                                         KeyRange range,
                                          Version* begin,
                                          Version end,
                                          int replyBufferSize,
                                          bool canReadPopped,
                                          ReadOptions readOptions,
-                                         bool encrypted,
-                                         Reference<ChangeFeedCacheData> cacheData,
-                                         Key tenantPrefix) {
+                                         bool encrypted) {
 	state std::vector<Future<Void>> fetchers(interfs.size());
 	state std::vector<Future<Void>> onErrors(interfs.size());
 	state std::vector<MutationAndVersionStream> streams(interfs.size());
@@ -10349,9 +10218,7 @@ ACTOR Future<Void> mergeChangeFeedStream(Reference<DatabaseContext> db,
 		                                      &tssDatas[i]);
 	}
 
-	wait(waitForAny(onErrors) ||
-	     mergeChangeFeedStreamInternal(
-	         results, rangeID, range, interfs, streams, begin, end, mergeCursorUID, db, cacheData, tenantPrefix));
+	wait(waitForAny(onErrors) || mergeChangeFeedStreamInternal(results, interfs, streams, begin, end, mergeCursorUID));
 
 	return Void();
 }
@@ -10403,10 +10270,7 @@ ACTOR Future<Void> singleChangeFeedStreamInternal(KeyRange range,
                                                   Key rangeID,
                                                   Version* begin,
                                                   Version end,
-                                                  Optional<ChangeFeedTSSValidationData>* tssData,
-                                                  Reference<DatabaseContext> db,
-                                                  Reference<ChangeFeedCacheData> cacheData,
-                                                  Key tenantPrefix) {
+                                                  Optional<ChangeFeedTSSValidationData>* tssData) {
 
 	state Promise<Void> refresh = results->refresh;
 	ASSERT(results->streams.size() == 1);
@@ -10462,7 +10326,6 @@ ACTOR Future<Void> singleChangeFeedStreamInternal(KeyRange range,
 				tssData->get().send(feedReply);
 			}
 
-			writeMutationsToCache(cacheData, db, feedReply.mutations, rangeID, range, tenantPrefix);
 			results->mutations.send(
 			    Standalone<VectorRef<MutationsAndVersionRef>>(feedReply.mutations, feedReply.arena));
 
@@ -10513,9 +10376,7 @@ ACTOR Future<Void> singleChangeFeedStream(Reference<DatabaseContext> db,
                                           int replyBufferSize,
                                           bool canReadPopped,
                                           ReadOptions readOptions,
-                                          bool encrypted,
-                                          Reference<ChangeFeedCacheData> cacheData,
-                                          Key tenantPrefix) {
+                                          bool encrypted) {
 	state Database cx(db);
 	state ChangeFeedStreamRequest req;
 	state Optional<ChangeFeedTSSValidationData> tssData;
@@ -10558,7 +10419,7 @@ ACTOR Future<Void> singleChangeFeedStream(Reference<DatabaseContext> db,
 	    req, interf.changeFeedStream, cx->enableLocalityLoadBalance ? &cx->queueModel : nullptr, &tssData);
 
 	wait(results->streams[0].onError() ||
-	     singleChangeFeedStreamInternal(range, results, rangeID, begin, end, &tssData, db, cacheData, tenantPrefix));
+	     singleChangeFeedStreamInternal(range, results, rangeID, begin, end, &tssData));
 
 	return Void();
 }
@@ -10605,50 +10466,6 @@ void coalesceChangeFeedLocations(std::vector<KeyRangeLocationInfo>& locations) {
 	locations = coalesced;
 }
 
-ACTOR Future<bool> getChangeFeedStreamFromDisk(Reference<DatabaseContext> db,
-                                               Reference<ChangeFeedData> results,
-                                               Key rangeID,
-                                               Version* begin,
-                                               Version end,
-                                               KeyRange range,
-                                               Key tenantPrefix) {
-	state bool foundEnd = false;
-	loop {
-		Key beginKey = changeFeedCacheKey(tenantPrefix, rangeID, range, *begin);
-		Key endKey = changeFeedCacheKey(tenantPrefix, rangeID, range, MAX_VERSION);
-		state RangeResult res = wait(db->storage->readRange(KeyRangeRef(beginKey, endKey),
-		                                                    CLIENT_KNOBS->CHANGE_FEED_CACHE_LIMIT_BYTES,
-		                                                    CLIENT_KNOBS->CHANGE_FEED_CACHE_LIMIT_BYTES));
-		state int idx = 0;
-
-		while (!foundEnd && idx < res.size()) {
-			Standalone<VectorRef<MutationsAndVersionRef>> mutations = decodeChangeFeedCacheValue(res[idx].value);
-			while (!mutations.empty() && mutations.front().version < *begin) {
-				mutations.pop_front();
-			}
-			while (!mutations.empty() && mutations.back().version >= end) {
-				mutations.pop_back();
-				foundEnd = true;
-			}
-			if (!mutations.empty()) {
-				*begin = mutations.back().version;
-				results->mutations.send(mutations);
-				wait(results->mutations.onEmpty());
-				wait(delay(0));
-				if (*begin > results->lastReturnedVersion.get()) {
-					results->lastReturnedVersion.set(*begin);
-				}
-			}
-			(*begin)++;
-			idx++;
-		}
-
-		if (foundEnd || !res.more) {
-			return foundEnd;
-		}
-	}
-}
-
 ACTOR Future<Void> getChangeFeedStreamActor(Reference<DatabaseContext> db,
                                             Reference<ChangeFeedData> results,
                                             Key rangeID,
@@ -10658,11 +10475,12 @@ ACTOR Future<Void> getChangeFeedStreamActor(Reference<DatabaseContext> db,
                                             int replyBufferSize,
                                             bool canReadPopped,
                                             ReadOptions readOptions,
-                                            bool encrypted,
-                                            Reference<ChangeFeedCacheData> cacheData,
-                                            Key tenantPrefix) {
+                                            bool encrypted) {
 	state Database cx(db);
 	state Span span("NAPI:GetChangeFeedStream"_loc);
+	db->usedAnyChangeFeeds = true;
+
+	results->endVersion = end;
 
 	state double sleepWithBackoff = CLIENT_KNOBS->WRONG_SHARD_SERVER_DELAY;
 	state Version lastBeginVersion = invalidVersion;
@@ -10755,15 +10573,12 @@ ACTOR Future<Void> getChangeFeedStreamActor(Reference<DatabaseContext> db,
 				                           interfs,
 				                           results,
 				                           rangeID,
-				                           range,
 				                           &begin,
 				                           end,
 				                           replyBufferSize,
 				                           canReadPopped,
 				                           readOptions,
-				                           encrypted,
-				                           cacheData,
-				                           tenantPrefix) ||
+				                           encrypted) ||
 				     cx->connectionFileChanged());
 			} else {
 				CODE_PROBE(true, "Change feed single cursor");
@@ -10778,9 +10593,7 @@ ACTOR Future<Void> getChangeFeedStreamActor(Reference<DatabaseContext> db,
 				                            replyBufferSize,
 				                            canReadPopped,
 				                            readOptions,
-				                            encrypted,
-				                            cacheData,
-				                            tenantPrefix) ||
+				                            encrypted) ||
 				     cx->connectionFileChanged());
 			}
 		} catch (Error& e) {
@@ -10843,93 +10656,6 @@ ACTOR Future<Void> getChangeFeedStreamActor(Reference<DatabaseContext> db,
 	}
 }
 
-ACTOR Future<Void> durableChangeFeedMonitor(Reference<DatabaseContext> db,
-                                            Reference<ChangeFeedData> results,
-                                            Key rangeID,
-                                            Version begin,
-                                            Version end,
-                                            KeyRange range,
-                                            int replyBufferSize,
-                                            bool canReadPopped,
-                                            ReadOptions readOptions,
-                                            bool encrypted,
-                                            Future<Key> tenantPrefix) {
-	state Optional<ChangeFeedCacheRange> cacheRange;
-	state Reference<ChangeFeedCacheData> data;
-	state Error err = success();
-	state Version originalBegin = begin;
-	results->endVersion = end;
-	db->usedAnyChangeFeeds = true;
-	try {
-		if (db->storage != nullptr) {
-			wait(db->initializeChangeFeedCache);
-			Key prefix = wait(tenantPrefix);
-			cacheRange = ChangeFeedCacheRange(prefix, rangeID, range);
-			if (db->changeFeedCaches.count(cacheRange.get())) {
-				auto cacheData = db->changeFeedCaches[cacheRange.get()];
-				if (begin < cacheData->popped) {
-					results->mutations.sendError(change_feed_popped());
-					return Void();
-				}
-				if (cacheData->version <= begin) {
-					bool foundEnd = wait(getChangeFeedStreamFromDisk(db, results, rangeID, &begin, end, range, prefix));
-					if (foundEnd) {
-						results->mutations.sendError(end_of_stream());
-						return Void();
-					}
-				}
-			}
-			if (end == MAX_VERSION) {
-				if (!db->changeFeedCaches.count(cacheRange.get())) {
-					data = makeReference<ChangeFeedCacheData>();
-					data->version = begin;
-					data->active = true;
-					db->changeFeedCaches[cacheRange.get()] = data;
-					db->rangeId_cacheData[cacheRange.get().rangeId][cacheRange.get()] = data;
-					Key durableFeedKey = changeFeedCacheFeedKey(cacheRange.get().tenantPrefix, rangeID, range);
-					Value durableFeedValue = changeFeedCacheFeedValue(begin, 0);
-					db->storage->set(KeyValueRef(durableFeedKey, durableFeedValue));
-				} else {
-					data = db->changeFeedCaches[cacheRange.get()];
-					if (!data->active && data->version <= begin) {
-						data->active = true;
-						if (originalBegin > data->latest + 1) {
-							data->version = originalBegin;
-							Key durableFeedKey = changeFeedCacheFeedKey(cacheRange.get().tenantPrefix, rangeID, range);
-							Value durableFeedValue = changeFeedCacheFeedValue(originalBegin, data->popped);
-							db->storage->set(KeyValueRef(durableFeedKey, durableFeedValue));
-						}
-					} else {
-						data = Reference<ChangeFeedCacheData>();
-					}
-				}
-			}
-		}
-		wait(getChangeFeedStreamActor(db,
-		                              results,
-		                              rangeID,
-		                              begin,
-		                              end,
-		                              range,
-		                              replyBufferSize,
-		                              canReadPopped,
-		                              readOptions,
-		                              encrypted,
-		                              data,
-		                              cacheRange.present() ? cacheRange.get().tenantPrefix : Key()));
-	} catch (Error& e) {
-		err = e;
-	}
-	if (data) {
-		data->active = false;
-		data->inactiveTime = now();
-	}
-	if (err.code() != error_code_success) {
-		throw err;
-	}
-	return Void();
-}
-
 Future<Void> DatabaseContext::getChangeFeedStream(Reference<ChangeFeedData> results,
                                                   Key rangeID,
                                                   Version begin,
@@ -10938,9 +10664,8 @@ Future<Void> DatabaseContext::getChangeFeedStream(Reference<ChangeFeedData> resu
                                                   int replyBufferSize,
                                                   bool canReadPopped,
                                                   ReadOptions readOptions,
-                                                  bool encrypted,
-                                                  Future<Key> tenantPrefix) {
-	return durableChangeFeedMonitor(Reference<DatabaseContext>::addRef(this),
+                                                  bool encrypted) {
+	return getChangeFeedStreamActor(Reference<DatabaseContext>::addRef(this),
 	                                results,
 	                                rangeID,
 	                                begin,
@@ -10949,8 +10674,7 @@ Future<Void> DatabaseContext::getChangeFeedStream(Reference<ChangeFeedData> resu
 	                                replyBufferSize,
 	                                canReadPopped,
 	                                readOptions,
-	                                encrypted,
-	                                tenantPrefix);
+	                                encrypted);
 }
 
 Version OverlappingChangeFeedsInfo::getFeedMetadataVersion(const KeyRangeRef& range) const {
@@ -11098,26 +10822,6 @@ ACTOR Future<Void> popChangeFeedMutationsActor(Reference<DatabaseContext> db, Ke
 	state Span span("NAPI:PopChangeFeedMutations"_loc);
 	db->usedAnyChangeFeeds = true;
 	++db->feedPops;
-
-	if (db->rangeId_cacheData.count(rangeID)) {
-		auto& feeds = db->rangeId_cacheData[rangeID];
-		for (auto& it : feeds) {
-			if (version > it.second->popped) {
-				it.second->popped = version;
-				Key beginKey = changeFeedCacheKey(it.first.tenantPrefix, it.first.rangeId, it.first.range, 0);
-				Key endKey = changeFeedCacheKey(it.first.tenantPrefix, it.first.rangeId, it.first.range, version);
-				db->storage->clear(KeyRangeRef(beginKey, endKey));
-				Key durableFeedKey = changeFeedCacheFeedKey(it.first.tenantPrefix, it.first.rangeId, it.first.range);
-				Value durableFeedValue = changeFeedCacheFeedValue(it.second->version, it.second->popped);
-				db->storage->set(KeyValueRef(durableFeedKey, durableFeedValue));
-				db->uncommittedCFBytes +=
-				    beginKey.size() + endKey.size() + durableFeedKey.size() + durableFeedValue.size();
-				if (db->uncommittedCFBytes > CLIENT_KNOBS->CHANGE_FEED_CACHE_FLUSH_BYTES) {
-					db->commitChangeFeedStorage->set(true);
-				}
-			}
-		}
-	}
 
 	state KeyRange keys = wait(getChangeFeedRange(db, cx, rangeID));
