+++ fdbclient/include/fdbclient/ClientKnobs.h
@@ -80,9 +80,6 @@ public:
 	int64_t CHANGE_FEED_STREAM_MIN_BYTES;
 	double CHANGE_FEED_START_INTERVAL;
 	bool CHANGE_FEED_COALESCE_LOCATIONS;
-	int64_t CHANGE_FEED_CACHE_FLUSH_BYTES;
-	double CHANGE_FEED_CACHE_EXPIRE_TIME;
-	int64_t CHANGE_FEED_CACHE_LIMIT_BYTES;
 
 	int MAX_BATCH_SIZE;
 	double GRV_BATCH_TIMEOUT;
