+++ fdbclient/include/fdbclient/DatabaseContext.h
@@ -38,7 +38,6 @@
 #include "fdbclient/CommitProxyInterface.h"
 #include "fdbclient/SpecialKeySpace.actor.h"
 #include "fdbclient/VersionVector.h"
-#include "fdbclient/IKeyValueStore.actor.h"
 #include "fdbrpc/QueueModel.h"
 #include "fdbrpc/MultiInterface.h"
 #include "flow/TDMetric.actor.h"
@@ -221,44 +220,6 @@ struct OverlappingChangeFeedsInfo {
 	Version getFeedMetadataVersion(const KeyRangeRef& feedRange) const;
 };
 
-struct ChangeFeedCacheRange {
-	Key tenantPrefix;
-	Key rangeId;
-	KeyRange range;
-
-	ChangeFeedCacheRange(Key tenantPrefix, Key rangeId, KeyRange range)
-	  : tenantPrefix(tenantPrefix), rangeId(rangeId), range(range) {}
-	ChangeFeedCacheRange(std::tuple<Key, Key, KeyRange> range)
-	  : tenantPrefix(std::get<0>(range)), rangeId(std::get<1>(range)), range(std::get<2>(range)) {}
-
-	bool operator==(const ChangeFeedCacheRange& rhs) const {
-		return tenantPrefix == rhs.tenantPrefix && rangeId == rhs.rangeId && range == rhs.range;
-	}
-};
-
-struct ChangeFeedCacheData : ReferenceCounted<ChangeFeedCacheData> {
-	Version version = -1; // The first version durably stored in the cache
-	Version latest =
-	    -1; // The last version durably store in the cache; this version will not be readable from disk before a commit
-	Version popped = -1; // The popped version of this change feed
-	bool active = false;
-	double inactiveTime = 0;
-};
-
-namespace std {
-template <>
-struct hash<ChangeFeedCacheRange> {
-	static constexpr std::hash<StringRef> hashFunc{};
-	std::size_t operator()(ChangeFeedCacheRange const& range) const {
-		std::size_t seed = 0;
-		boost::hash_combine(seed, hashFunc(range.rangeId));
-		boost::hash_combine(seed, hashFunc(range.range.begin));
-		boost::hash_combine(seed, hashFunc(range.range.end));
-		return seed;
-	}
-};
-} // namespace std
-
 class DatabaseContext : public ReferenceCounted<DatabaseContext>, public FastAllocated<DatabaseContext>, NonCopyable {
 public:
 	static DatabaseContext* allocateOnForeignThread() {
@@ -439,8 +400,7 @@ public:
 	                                 int replyBufferSize = -1,
 	                                 bool canReadPopped = true,
 	                                 ReadOptions readOptions = { ReadType::NORMAL, CacheResult::False },
-	                                 bool encrypted = false,
-	                                 Future<Key> tenantPrefix = Key());
+	                                 bool encrypted = false);
 
 	Future<OverlappingChangeFeedsInfo> getOverlappingChangeFeeds(KeyRangeRef ranges, Version minVersion);
 	Future<Void> popChangeFeedMutations(Key rangeID, Version version);
@@ -556,17 +516,6 @@ public:
 	std::unordered_map<Key, KeyRange> changeFeedCache;
 	std::unordered_map<UID, ChangeFeedStorageData*> changeFeedUpdaters;
 	std::map<UID, ChangeFeedData*> notAtLatestChangeFeeds;
-
-	IKeyValueStore* storage = nullptr;
-	Future<Void> changeFeedStorageCommitter;
-	Future<Void> initializeChangeFeedCache = Void();
-	int64_t uncommittedCFBytes = 0;
-	Reference<AsyncVar<bool>> commitChangeFeedStorage;
-
-	std::unordered_map<ChangeFeedCacheRange, Reference<ChangeFeedCacheData>> changeFeedCaches;
-	std::unordered_map<Key, std::unordered_map<ChangeFeedCacheRange, Reference<ChangeFeedCacheData>>> rangeId_cacheData;
-
-	void setStorage(IKeyValueStore* storage);
 
 	Reference<ChangeFeedStorageData> getStorageData(StorageServerInterface interf);
 	Version getMinimumChangeFeedVersion();
