+++ fdbserver/KeyValueStoreShardedRocksDB.actor.cpp
@@ -38,7 +38,6 @@
 
 #endif // SSD_ROCKSDB_EXPERIMENTAL
 
-#include "fdbserver/Knobs.h"
 #include "fdbserver/IKeyValueStore.h"
 #include "fdbserver/RocksDBCheckpointUtils.actor.h"
 #include "flow/actorcompiler.h" // has to be last include
