+++ fdbserver/RemoteIKeyValueStore.actor.cpp
@@ -19,7 +19,6 @@
  */
 
 #include "fdbserver/IKeyValueStore.h"
-#include "fdbserver/ServerDBInfo.actor.h"
 #include "flow/ActorCollection.h"
 #include "flow/Error.h"
 #include "flow/Platform.h"
