+++ fdbserver/include/fdbserver/IPager.h
@@ -27,7 +27,7 @@
 #include "fdbclient/FDBTypes.h"
 #include "fdbclient/GetEncryptCipherKeys.h"
 #include "fdbclient/Tenant.h"
-#include "fdbclient/IClosable.h"
+#include "fdbserver/IClosable.h"
 #include "flow/EncryptUtils.h"
 #include "flow/Error.h"
 #include "flow/FastAlloc.h"
