+++ fdbserver/include/fdbserver/WorkerInterface.actor.h
@@ -891,20 +891,6 @@ struct InitializeBlobWorkerRequest {
 	UID interfaceId;
 	KeyValueStoreType storeType;
 	ReplyPromise<InitializeBlobWorkerReply> reply;
-	EncryptionAtRestMode encryptMode;
-
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, reqId, interfaceId, storeType, reply, encryptMode);
-	}
-};
-
-struct InitializeBlobWorkerRequestOld {
-	constexpr static FileIdentifier file_identifier = 5838547;
-	UID reqId;
-	UID interfaceId;
-	KeyValueStoreType storeType;
-	ReplyPromise<InitializeBlobWorkerReply> reply;
 
 	template <class Ar>
 	void serialize(Ar& ar) {
