+++ fdbserver/worker.actor.cpp
@@ -1778,20 +1778,6 @@ ACTOR Future<Void> updateClusterId(UID ccClusterId, Reference<AsyncVar<Optional<
 	return Void();
 }
 
-ACTOR Future<Void> deleteStorageFile(KeyValueStoreType storeType,
-                                     std::string filename,
-                                     UID storeID,
-                                     int64_t memoryLimit,
-                                     Reference<AsyncVar<ServerDBInfo>> dbInfo) {
-	state IKeyValueStore* kvs = openKVStore(storeType, filename, storeID, memoryLimit, false, false, false, dbInfo, {});
-	wait(ready(kvs->init()));
-	TraceEvent("KVSRemoved").detail("Reason", "WorkerRemoved");
-	kvs->dispose();
-	CODE_PROBE(true, "Removed stale disk file");
-	TraceEvent("RemoveStorageDisk").detail("Filename", filename).detail("StoreID", storeID);
-	return Void();
-}
-
 ACTOR Future<Void> cleanupStaleStorageDisk(Reference<AsyncVar<ServerDBInfo>> dbInfo,
                                            std::unordered_map<UID, StorageDiskCleaner>* cleaners,
                                            UID storeID,
@@ -1821,7 +1807,12 @@ ACTOR Future<Void> cleanupStaleStorageDisk(Reference<AsyncVar<ServerDBInfo>> dbI
 			if (e.code() == error_code_worker_removed) {
 				// delete the files on disk
 				if (fileExists(cleaner.filename)) {
-					wait(deleteStorageFile(cleaner.storeType, cleaner.filename, storeID, memoryLimit, dbInfo));
+					state IKeyValueStore* kvs = openKVStore(
+					    cleaner.storeType, cleaner.filename, storeID, memoryLimit, false, false, false, dbInfo, {});
+					wait(ready(kvs->init()));
+					kvs->dispose();
+					CODE_PROBE(true, "Removed stale disk file");
+					TraceEvent("RemoveStorageDisk").detail("Filename", cleaner.filename).detail("StoreID", storeID);
 				}
 
 				// remove the cleaner
@@ -2146,7 +2137,7 @@ ACTOR Future<Void> workerServer(Reference<IClusterConnectionRecord> connRecord,
 				logData.back().uid = s.storeID;
 				errorForwarders.add(forwardError(errors, Role::SHARED_TRANSACTION_LOG, s.storeID, tl));
 			} else if (s.storedComponent == DiskStore::BlobWorker) {
-				if (blobWorkerFuture.isReady() && SERVER_KNOBS->BLOB_WORKER_DISK_ENABLED) {
+				if (blobWorkerFuture.isReady()) {
 					LocalLineage _;
 					getCurrentLineage()->modify(&RoleLineage::role) = ProcessClass::ClusterRole::BlobWorker;
 
@@ -2174,8 +2165,7 @@ ACTOR Future<Void> workerServer(Reference<IClusterConnectionRecord> connRecord,
 					                                   false,
 					                                   false,
 					                                   dbInfo,
-					                                   Optional<EncryptionAtRestMode>(),
-					                                   FLOW_KNOBS->BLOB_WORKER_PAGE_CACHE);
+					                                   EncryptionAtRestMode());
 					filesClosed.add(data->onClosed());
 
 					Promise<Void> recovery;
@@ -2186,7 +2176,16 @@ ACTOR Future<Void> workerServer(Reference<IClusterConnectionRecord> connRecord,
 					errorForwarders.add(forwardError(errors, Role::BLOB_WORKER, recruited.id(), bw));
 				} else {
 					CODE_PROBE(true, "Multiple blob workers after reboot", probe::decoration::rare);
-					recoveries.push_back(deleteStorageFile(s.storeType, s.filename, s.storeID, memoryLimit, dbInfo));
+					IKeyValueStore* data = openKVStore(s.storeType,
+					                                   s.filename,
+					                                   UID(),
+					                                   memoryLimit,
+					                                   false,
+					                                   false,
+					                                   false,
+					                                   dbInfo,
+					                                   EncryptionAtRestMode());
+					data->dispose();
 				}
 			}
 		}
@@ -2846,8 +2845,7 @@ ACTOR Future<Void> workerServer(Reference<IClusterConnectionRecord> connRecord,
 						                   false,
 						                   false,
 						                   dbInfo,
-						                   req.encryptMode,
-						                   FLOW_KNOBS->BLOB_WORKER_PAGE_CACHE);
+						                   EncryptionAtRestMode());
 						filesClosed.add(data->onClosed());
 					}
